requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
aiosqlite
asyncpg
bcrypt
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
jinja2
psycopg2-binary
python-multipart
